requirement.txt
amqp                   5.3.1
asgiref                3.8.1
billiard               4.2.1
celery                 5.5.3
click                  8.2.1
click-didyoumean       0.3.1
click-plugins          1.1.1.2
click-repl             0.3.0
colorama               0.4.6
Django                 5.2.3
django-cors-headers    4.7.0
django-environ         0.12.0
djangorestframework    3.16.0
drf-yasg               1.21.10
faker
inflection             0.5.1
kombu                  5.5.4
mysql-connector-python 9.3.0
mysqlclient            2.2.7
packaging              25.0
pip                    25.1.1
prompt_toolkit         3.0.51
python-dateutil        2.9.0.post0
python-dotenv          1.1.1
pytz                   2025.2
PyYAML                 6.0.2
setuptools             80.9.0
six                    1.17.0
sqlparse               0.5.3
tzdata                 2025.2
uritemplate            4.2.0
vine                   5.1.0
wcwidth                0.2.13
erlangOTP              28.0.1
